minor styling + prev/next day buttons fix
