+++ index.html
@@ -1,1 +1,14 @@
-<!DOCTYPE html><html lang="en"><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1"><link rel="shortcut icon" href="/calendar-exercise/favicon.ico"><title>Daily Agenda</title><link href="/calendar-exercise/static/css/main.92053648.css" rel="stylesheet"></head><body><div id="root"></div><script type="text/javascript" src="/calendar-exercise/static/js/main.cc307f24.js"></script></body></html>
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width,initial-scale=1">
+    <link rel="shortcut icon" href="/calendar-exercise/favicon.ico">
+    <title>Daily Agenda</title>
+    <link href="/calendar-exercise/static/css/main.92053648.css" rel="stylesheet">
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="text/javascript" src="/calendar-exercise/static/js/main.cc307f24.js"></script>
+  </body>
+</html>
