+++ src/components/Calendar.css
@@ -1,6 +1,4 @@
-/*
- * TODO: Calendar should have rounded corners
- */
+
 .calendar {
     border: 1px solid #666;
     border-radius: 5px;
