+++ src/components/EventDetailOverlay.css
@@ -1,6 +1,3 @@
-/*
- * TODO: Overlay container should have rounded corners
- */
 
 .event-detail-overlay {
     position: fixed;
