+++ src/components/Page.jsx
@@ -1,0 +1,96 @@
+import React from 'react';
+import Calendar from './Calendar.jsx';
+import EventDetailOverlay from './EventDetailOverlay.jsx';
+import {filterEventsByDay, getEventFromEvents, getDisplayDate} from '../utils';
+import DATA_SET from '../utils/data';
+
+import './Page.css';
+
+const DayNavigator = ({dateDisplay, onPrev, onNext}) => {
+  return (
+    <nav className="page__nav">
+      <button
+        className="page__nav-button page__prev-day"
+        title="Go to previous day"
+        onClick={onPrev}
+      />
+      <h2 className="page__date">{dateDisplay}</h2>
+      <button
+        className="page__nav-button page__next-day"
+        title="Go to next day"
+        onClick={onNext}
+      />
+    </nav>
+  );
+};
+
+export default class Page extends React.PureComponent {
+  state = {
+    // unfiltered list of events
+    events: DATA_SET,
+
+    // The currently selected day represented by numerical timestamp
+    day: Date.now(),
+
+    // The currently selected event in the agenda
+    // (mainly to trigger event detail overlay)
+    selectedEventId: undefined
+  }
+
+  _handleSelectEvent(selectedEventId) {
+    this.setState({selectedEventId});
+  }
+
+  _handleEventDetailOverlayClose() {
+    this.setState({selectedEventId: undefined});
+  }
+
+  _handlePrev() {
+    // change events too, look at fake data
+    this.setState((prevState) => {
+      console.log(prevState)
+      return {
+        day: prevState.day - (24*60*60*1000)
+      }
+    })
+  }
+
+  _handleNext() {
+    this.setState((prevState) => {
+      return {
+        day: prevState.day + (24*60*60*1000)
+      }
+    })
+  }
+
+  render() {
+    let {events, day, selectedEventId} = this.state;
+    let filteredEvents = filterEventsByDay(events, day);
+    let selectedEvent = getEventFromEvents(events, selectedEventId);
+    let eventDetailOverlay;
+
+    if (selectedEvent) {
+      eventDetailOverlay = (
+          <EventDetailOverlay
+              event={selectedEvent}
+              onClose={this._handleEventDetailOverlayClose.bind(this)}
+          />
+      );
+    }
+
+    return (
+      <div className="page">
+        <header className="page__header">
+            <h1 className="page__title">Daily Agenda</h1>
+        </header>
+        <DayNavigator
+            dateDisplay={getDisplayDate(day)}
+            onPrev={this._handlePrev.bind(this)}
+            onNext={this._handleNext.bind(this)}
+        />
+        <Calendar events={filteredEvents} onSelectEvent={this._handleSelectEvent.bind(this)} />
+        {eventDetailOverlay}
+      </div>
+    );
+  }
+}
