+++ src/components/TimeSlotEvent.css
@@ -1,5 +1,4 @@
 /*
- * TODO: Time slot events should have rounded corners
  * TODO: Past events should display faded-out
  */
 .time-slot-event {
