+++ src/index.css
@@ -14,5 +14,4 @@ body {
     color: #212121;
     font-family: "Helvetica Neue", "Calibri Light", Roboto, sans-serif;
     letter-spacing: 0.02em;
-    background: linear-gradient(white,lightgray);
 }
